only owner can delete their tasks

+++ src/Controller/TaskController.php
@@ -51,7 +51,7 @@ class TaskController extends AbstractController
 
         $form->handleRequest($request);
 
-        if ($form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
@@ -83,11 +83,15 @@ class TaskController extends AbstractController
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        if($task->getUser() != $this->getUser()){
+            $this->addFlash("error", "Vous n'êtes pas l'auteur de cette tâche !");
+        }else{
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+            $this->addFlash('success', 'La tâche à bien été supprimée.');
+        }
 
         return $this->redirectToRoute('task_list');
     }
